Extract helpers from CodeViewer component
